+++ exts/wandelbots.omni/wandelbots/omni/usd/tcp_utils.py
@@ -7,10 +7,7 @@ import omni.kit
 from omni.kit.property.usd import prim_selection_payload
 import omni.kit.commands
 
-try:
-    import isaacsim.core.utils.stage as stage_utils
-except ImportError:
-    import omni.isaac.core.utils.stage as stage_utils  # type: ignore
+import isaacsim.core.utils.stage as stage_utils
 
 
 class TcpUtils: